Revert "Extract stable `stwo-types` crate containing field and struct types used downstream (#1331)"
This reverts commit 95932933425a5fb07fa0bcae55bc8d850abe5173.
