+++ Cargo.toml
@@ -6,7 +6,6 @@ members = [
     "crates/constraint-framework",
     "crates/examples",
     "crates/std-shims",
-    "crates/stwo-types",
 ]
 exclude = ["ensure-verifier-no_std"]
 resolver = "2"
@@ -33,7 +32,6 @@ rayon = { version = "1.10.0", optional = false }
 rand = { version = "0.8.5", default-features = false, features = ["small_rng"] }
 serde = { version = "1.0", default-features = false, features = ["derive"] }
 hashbrown = { version = ">=0.15.2", features = ["serde"] }
-stwo-types = { path = "crates/stwo-types", version = "2.0.1", default-features = false }
 
 [profile.bench]
 codegen-units = 1
