--- crates/stwo-types/Cargo.toml
@@ -1,27 +1,0 @@
-[package]
-name = "stwo-types"
-version = "2.0.1"
-edition.workspace = true
-license.workspace = true
-description = "Core field types shared across stwo crates"
-
-[features]
-default = ["std"]
-std = ["serde/std", "num-traits/std", "rand/std", "std-shims/std"]
-parallel = ["dep:rayon"]
-
-[dependencies]
-bytemuck = { workspace = true, features = ["derive", "extern_crate_alloc"] }
-num-traits.workspace = true
-rand.workspace = true
-serde.workspace = true
-rayon = { workspace = true, optional = true }
-std-shims = { path = "../std-shims", version = "1.0.0", default-features = false }
-
-[lints.rust]
-future-incompatible = "deny"
-nonstandard-style = "deny"
-rust-2018-idioms = "deny"
-
-[lints.clippy]
-missing_const_for_fn = "warn"
