--- crates/stwo-types/src/lib.rs
@@ -1,9 +1,0 @@
-#![cfg_attr(not(feature = "std"), no_std)]
-
-pub mod fields;
-mod utils;
-
-pub use fields::*;
-#[cfg(test)]
-pub use fields::{m31, qm31};
-pub use utils::uninit_vec;
